test: getOtherUserInfo 테스트 작성 #67
